message코드정리
message코드정리

+++ gigsforgeeks/src/com/gigsforgeeks/message/controller/MessageDeleteServlet.java
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import com.gigsforgeeks.member.model.vo.Member;
 import com.gigsforgeeks.message.model.service.MessageService;
 
 /**
@@ -31,19 +32,29 @@ public class MessageDeleteServlet extends HttpServlet {
 	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		int messageNo = Integer.parseInt(request.getParameter("nno"));
+		HttpSession session = request.getSession();
+		Member loginUser = (Member)session.getAttribute("loginUser");
 		
-		int result = new MessageService().deleteMessage(messageNo);
-		if(result > 0) {	
-			HttpSession session = request.getSession();
-			session.setAttribute("alertMsg", "메세지를 삭제하였습니다.");
-			response.sendRedirect(request.getContextPath()+"/list.ms?currentPage=1");
-		}else {
-			request.setAttribute("errorMsg", "메세지삭제를 실패하였습니다.");
+		if(loginUser != null) { // 현재 사용자가 회원인 경우
 			
-			RequestDispatcher view = request.getRequestDispatcher("views/common/errorPage.jsp");
+			int messageNo = Integer.parseInt(request.getParameter("nno"));
+			
+			int result = new MessageService().deleteMessage(messageNo);
+			if(result > 0) {	
+				
+				session.setAttribute("alertMsg", "메세지를 삭제하였습니다.");
+				response.sendRedirect(request.getContextPath()+"/list.ms?currentPage=1");
+			}else {
+				request.setAttribute("errorMsg", "메세지삭제를 실패하였습니다.");
+				
+				RequestDispatcher view = request.getRequestDispatcher("views/common/errorPage.jsp");
+				view.forward(request, response);
+			}
+		}else { // 비회원인 경우
+			request.getSession().setAttribute("alertMsg", "로그인 후에 이용하세요.");
+			RequestDispatcher view = request.getRequestDispatcher("views/member/login.jsp");
 			view.forward(request, response);
-		}
+			}
